src/backend/recommendation_service/requirements.txt
implicit~=0.7.2
kafka-python
numpy
pandas~=2.2.2
scipy~=1.14.0
fastapi[standard]~=0.112.0
python-dotenv~=1.0.1
redis[hiredis]
Flask
waitress